@@ -6,59 +6,57 @@ using IpfsShipyard.Ipfs.Core;
 using IpfsShipyard.Ipfs.Core.CoreApi;
 using Newtonsoft.Json.Linq;
 
-namespace IpfsShipyard.Ipfs.Http.CoreApi
+namespace IpfsShipyard.Ipfs.Http.CoreApi;
+
+class BootstrapApi : IBootstrapApi
 {
-    class BootstrapApi : IBootstrapApi
+    IpfsClient _ipfs;
+
+    internal BootstrapApi(IpfsClient ipfs)
     {
-        IpfsClient _ipfs;
-
-        internal BootstrapApi(IpfsClient ipfs)
-        {
-            _ipfs = ipfs;
-        }
-
-        public async Task<MultiAddress> AddAsync(MultiAddress address, CancellationToken cancel = default(CancellationToken))
-        {
-            var json = await _ipfs.DoCommandAsync("bootstrap/add", cancel, address.ToString());
-            var addrs = (JArray)(JObject.Parse(json)["Peers"]);
-            var a = addrs.FirstOrDefault();
-            if (a == null)
-                return null;
-            return new MultiAddress((string)a);
-        }
-
-        public async Task<IEnumerable<MultiAddress>> AddDefaultsAsync(CancellationToken cancel = default(CancellationToken))
-        {
-            var json = await _ipfs.DoCommandAsync("bootstrap/add/default", cancel);
-            var addrs = (JArray)(JObject.Parse(json)["Peers"]);
-            return addrs
-                .Select(a => MultiAddress.TryCreate((string)a))
-                .Where(ma => ma != null);
-        }
-
-        public async Task<IEnumerable<MultiAddress>> ListAsync(CancellationToken cancel = default(CancellationToken))
-        {
-            var json = await _ipfs.DoCommandAsync("bootstrap/list", cancel);
-            var addrs = (JArray)(JObject.Parse(json)["Peers"]);
-            return addrs
-                .Select(a => MultiAddress.TryCreate((string)a))
-                .Where(ma => ma != null);
-        }
-
-        public Task RemoveAllAsync(CancellationToken cancel = default(CancellationToken))
-        {
-            return _ipfs.DoCommandAsync("bootstrap/rm/all", cancel);
-        }
-
-        public async Task<MultiAddress> RemoveAsync(MultiAddress address, CancellationToken cancel = default(CancellationToken))
-        {
-            var json = await _ipfs.DoCommandAsync("bootstrap/rm", cancel, address.ToString());
-            var addrs = (JArray)(JObject.Parse(json)["Peers"]);
-            var a = addrs.FirstOrDefault();
-            if (a == null)
-                return null;
-            return new MultiAddress((string)a);
-        }
+        _ipfs = ipfs;
     }
 
+    public async Task<MultiAddress> AddAsync(MultiAddress address, CancellationToken cancel = default(CancellationToken))
+    {
+        var json = await _ipfs.DoCommandAsync("bootstrap/add", cancel, address.ToString());
+        var addrs = (JArray)(JObject.Parse(json)["Peers"]);
+        var a = addrs.FirstOrDefault();
+        if (a == null)
+            return null;
+        return new MultiAddress((string)a);
+    }
+
+    public async Task<IEnumerable<MultiAddress>> AddDefaultsAsync(CancellationToken cancel = default(CancellationToken))
+    {
+        var json = await _ipfs.DoCommandAsync("bootstrap/add/default", cancel);
+        var addrs = (JArray)(JObject.Parse(json)["Peers"]);
+        return addrs
+            .Select(a => MultiAddress.TryCreate((string)a))
+            .Where(ma => ma != null);
+    }
+
+    public async Task<IEnumerable<MultiAddress>> ListAsync(CancellationToken cancel = default(CancellationToken))
+    {
+        var json = await _ipfs.DoCommandAsync("bootstrap/list", cancel);
+        var addrs = (JArray)(JObject.Parse(json)["Peers"]);
+        return addrs
+            .Select(a => MultiAddress.TryCreate((string)a))
+            .Where(ma => ma != null);
+    }
+
+    public Task RemoveAllAsync(CancellationToken cancel = default(CancellationToken))
+    {
+        return _ipfs.DoCommandAsync("bootstrap/rm/all", cancel);
+    }
+
+    public async Task<MultiAddress> RemoveAsync(MultiAddress address, CancellationToken cancel = default(CancellationToken))
+    {
+        var json = await _ipfs.DoCommandAsync("bootstrap/rm", cancel, address.ToString());
+        var addrs = (JArray)(JObject.Parse(json)["Peers"]);
+        var a = addrs.FirstOrDefault();
+        if (a == null)
+            return null;
+        return new MultiAddress((string)a);
+    }
 }
