@@ -3,24 +3,23 @@ using System.Threading.Tasks;
 using IpfsShipyard.Ipfs.Core.CoreApi;
 using Newtonsoft.Json.Linq;
 
-namespace IpfsShipyard.Ipfs.Http.CoreApi
+namespace IpfsShipyard.Ipfs.Http.CoreApi;
+
+class DnsApi : IDnsApi
 {
-    class DnsApi : IDnsApi
+    private IpfsClient _ipfs;
+
+    internal DnsApi(IpfsClient ipfs)
     {
-        private IpfsClient _ipfs;
+        _ipfs = ipfs;
+    }
 
-        internal DnsApi(IpfsClient ipfs)
-        {
-            _ipfs = ipfs;
-        }
-
-        public async Task<string> ResolveAsync(string name, bool recursive = false, CancellationToken cancel = default(CancellationToken))
-        {
-            var json = await _ipfs.DoCommandAsync("dns", cancel,
-                name,
-                $"recursive={recursive.ToString().ToLowerInvariant()}");
-            var path = (string)(JObject.Parse(json)["Path"]);
-            return path;
-        }
+    public async Task<string> ResolveAsync(string name, bool recursive = false, CancellationToken cancel = default(CancellationToken))
+    {
+        var json = await _ipfs.DoCommandAsync("dns", cancel,
+            name,
+            $"recursive={recursive.ToString().ToLowerInvariant()}");
+        var path = (string)(JObject.Parse(json)["Path"]);
+        return path;
     }
 }
