@@ -6,42 +6,40 @@ using IpfsShipyard.Ipfs.Core;
 using IpfsShipyard.Ipfs.Core.CoreApi;
 using Newtonsoft.Json.Linq;
 
-namespace IpfsShipyard.Ipfs.Http.CoreApi
+namespace IpfsShipyard.Ipfs.Http.CoreApi;
+
+class PinApi : IPinApi
 {
-    class PinApi : IPinApi
+    private IpfsClient _ipfs;
+
+    internal PinApi(IpfsClient ipfs)
     {
-        private IpfsClient _ipfs;
+        _ipfs = ipfs;
+    }
 
-        internal PinApi(IpfsClient ipfs)
-        {
-            _ipfs = ipfs;
-        }
+    public async Task<IEnumerable<Cid>> AddAsync(string path, bool recursive = true, CancellationToken cancel = default(CancellationToken))
+    {
+        var opts = "recursive=" + recursive.ToString().ToLowerInvariant();
+        var json = await _ipfs.DoCommandAsync("pin/add", cancel, path, opts);
+        return ((JArray)JObject.Parse(json)["Pins"])
+            .Select(p => (Cid)(string)p);
+    }
 
-        public async Task<IEnumerable<Cid>> AddAsync(string path, bool recursive = true, CancellationToken cancel = default(CancellationToken))
-        {
-            var opts = "recursive=" + recursive.ToString().ToLowerInvariant();
-            var json = await _ipfs.DoCommandAsync("pin/add", cancel, path, opts);
-            return ((JArray)JObject.Parse(json)["Pins"])
-                .Select(p => (Cid)(string)p);
-        }
+    public async Task<IEnumerable<Cid>> ListAsync(CancellationToken cancel = default(CancellationToken))
+    {
+        var json = await _ipfs.DoCommandAsync("pin/ls", cancel);
+        var keys = (JObject)(JObject.Parse(json)["Keys"]);
+        return keys
+            .Properties()
+            .Select(p => (Cid)p.Name);
+    }
 
-        public async Task<IEnumerable<Cid>> ListAsync(CancellationToken cancel = default(CancellationToken))
-        {
-            var json = await _ipfs.DoCommandAsync("pin/ls", cancel);
-            var keys = (JObject)(JObject.Parse(json)["Keys"]);
-            return keys
-                .Properties()
-                .Select(p => (Cid)p.Name);
-        }
-
-        public async Task<IEnumerable<Cid>> RemoveAsync(Cid id, bool recursive = true, CancellationToken cancel = default(CancellationToken))
-        {
-            var opts = "recursive=" + recursive.ToString().ToLowerInvariant();
-            var json = await _ipfs.DoCommandAsync("pin/rm", cancel, id, opts);
-            return ((JArray)JObject.Parse(json)["Pins"])
-                .Select(p => (Cid)(string)p);
-        }
-
+    public async Task<IEnumerable<Cid>> RemoveAsync(Cid id, bool recursive = true, CancellationToken cancel = default(CancellationToken))
+    {
+        var opts = "recursive=" + recursive.ToString().ToLowerInvariant();
+        var json = await _ipfs.DoCommandAsync("pin/rm", cancel, id, opts);
+        return ((JArray)JObject.Parse(json)["Pins"])
+            .Select(p => (Cid)(string)p);
     }
 
 }
